Add chant for SimpleHTTPServer
Why oh why doesn't Ruby ship with this!

For retrieving data over a network the browser
will use the protocol that the original page was
loaded with. In the case of a "double click" on
the file that would be `file:///` which muddles up
AJAX. You want to serve via `http://`.